Mozilla bug link replaced with the more exact one
The Mozilla bug referenced in the text is not about absolute positioning _in_ the cells, but about relative positioning of _the cells themselves._ Shouldn't this reference be replaced with the link to the bug that is specifically about positioning _in_ the cells?

Also, both these Mozilla bugs have been fixed in 2014. Shouldn't it be mentioned in the text?